pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cross-tx-pma-detector"
version = "0.1.0"
description = "Cross-Transaction Price Manipulation Attack Detection using Temporal GNN"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torch-geometric>=2.4",
    "torch-geometric-temporal>=0.54",
    "web3>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "networkx>=3.0",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
    "requests>=2.31",
    "tqdm>=4.65",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "wandb>=0.15",
    "ipykernel>=6.0",
    "jupyter>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
